fixed a bug where raw images weren't being flipwed for slabs 4-6

+++ reconstruction/nonlinear_2d_alignment.py
@@ -134,9 +134,10 @@ def receptor_2d_alignment( df, rec_fn, srv_fn, mv_dir, output_dir, resolution, r
     os.makedirs(tfm_dir,exist_ok=True)
 
     num_cores = min(14, multiprocessing.cpu_count() )
-    print('num cores', num_cores)
+    
     to_do_df = pd.DataFrame([])
     to_do_resample_df = pd.DataFrame([])
+
     df['tfm_affine']=['']*df.shape[0]
 
     for i, row in df.iterrows() :
@@ -170,9 +171,7 @@ def receptor_2d_alignment( df, rec_fn, srv_fn, mv_dir, output_dir, resolution, r
         df['tfm'].loc[ df['slab_order'] == y ] = tfm_fn
         df['tfm_affine'].loc[ df['slab_order'] == y ] = tfm_affine_fn
 
-    #df['init_tfm'] = df['tfm']
-
-    return df
+        return df
 
 def concatenate_sections_to_volume(df, rec_fn, output_dir, out_fn, target_str='rsl'):
     exit_flag=False
@@ -191,15 +190,20 @@ def concatenate_sections_to_volume(df, rec_fn, output_dir, out_fn, target_str='r
         fn = f'{tfm_dir}/y-{y}_{target_str}.nii.gz' 
         
         df[target_name].loc[i] = fn
-        try : 
-            out_vol[:,int(y),:] = nib.load(fn).get_fdata()
-        except EOFError :
-            print('Error:', fn)
-            os.remove(fn)
-            exit_flag=True
 
-        if exit_flag : exit(1)
-    print('\t\tWriting 3D non-linear:', out_fn)
-    #out_vol = np.flip(out_vol, axis=1)
-    nib.Nifti1Image(out_vol, hires_img.affine).to_filename(out_fn)
+    if not os.path.exists(out_fn) :
+        for idx, (i, row) in enumerate(df.iterrows()):
+            fn = df[target_name].loc[i]
+            y = row['slab_order'] 
+
+            try : 
+                out_vol[:,int(y),:] = nib.load(fn).get_fdata()
+            except EOFError :
+                print('Error:', fn)
+                os.remove(fn)
+                exit_flag=True
+
+            if exit_flag : exit(1)
+        print('\t\tWriting 3D non-linear:', out_fn)
+        nib.Nifti1Image(out_vol, hires_img.affine).to_filename(out_fn)
     return df
